+++ packages/polaris-viz/src/components/SparkLineChart/SparkLineChart.tsx
@@ -39,7 +39,12 @@ export function SparkLineChart(props: SparkLineChartProps) {
       sparkChart
     >
       {state !== ChartState.Success ? (
-        <ChartSkeleton type="Spark" state={state} errorText={errorText} />
+        <ChartSkeleton
+          type="Spark"
+          state={state}
+          errorText={errorText}
+          theme={theme}
+        />
       ) : (
         <Chart
           data={data}