+++ packages/polaris-viz/CHANGELOG.md
@@ -11,6 +11,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Fixed horizontal bar charts not animating their bars in on mount.
 - Fixed links from `polaris-viz.shopify.io` to `polaris-viz.shopify.com` in documentation.
+- Fixed `theme` not being passed to `<ChartSkeleton />`.
 
 ## [7.0.0] - 2022-08-12